enhance testing method: add more level 2 nodes under multiple level by selecting a wanted count
